requirements.txt
flask
python-binance==1.0.28